Added qr generator module

+++ custom_modules/FileOperator.py
@@ -1,5 +1,6 @@
 import os
 from custom_modules.FileValidator import fileExists, isFile
+from custom_modules.TypeTester import arg_is_a_list
 
 
 def delete_file(file_path):
@@ -9,19 +10,20 @@ def delete_file(file_path):
 
 
 def append_file(file_path, list_data):
-    if fileExists(file_path) and isFile(file_path):
-        deleted = delete_file(file_path)
+    if arg_is_a_list(list_data):
+        if fileExists(file_path) and isFile(file_path):
+            deleted = delete_file(file_path)
 
-        if deleted:
+            if deleted:
+                with open(file_path, "a") as f:
+                    for d in list_data:
+                        f.write(d)
+
+                return fileExists(file_path)
+        else:
             with open(file_path, "a") as f:
                 for d in list_data:
                     f.write(d)
 
             return fileExists(file_path)
-    else:
-        with open(file_path, "a") as f:
-            for d in list_data:
-                f.write(d)
-
-        return fileExists(file_path)
     return None
